refactor: combined component types into one file
